requirements.txt
pywin32==306
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
cast-unknown==0.1.7
numpy==1.25.2
pillow==10.0.0
mouse==0.7.1
adb-shell==0.4.4
